Добавлено удаление

+++ GoalsTrackerApp/Managers/DataManager+GoalRecord.swift
@@ -92,7 +92,13 @@ extension DataManager: GoalRecordDataManagerProtocol {
     func deleteGoalRecord(for goalID: UUID, date: Date) {
         context.performAndWait {
             guard let goalRecord = readGoalRecord(for: goalID, date: date) else { return }
-
+            
+            while let goalRecord = readGoalRecord(for: goalID, date: date) {
+                context.delete(goalRecord)
+            }
+            
+            guard context.hasChanges else { return }
+            
             do {
                 context.delete(goalRecord)
                 try context.save()
